chore: rename referral folder
